Add PureCollider option for rigid bodies
Pure colliders allow rigid bodies to be instantiated for collision detection, while allowing
other rigid bodies to pass through them and without them exerting any forces on any
other rigid body.  Effectively it allows you to use rigid bodies as fast collision zones.

For example, if you wanted to know when another rigid body is within a certain distance of
an object, where you may have many such objects in the scene, you could create a rigid
body with a sphere shape for each of those objects, and set the sphere rigid body as a
pure collider.  When the CollisionBegin event is raised in the world with a sphere rigid
body as one of the arguments, you know that the other object is now within range.